Dashboard_grafana/Gateway/requirements.txt
pyserial
influxdb-client
python-dotenv